Move abstract_gas_cell to a static method of KEVM
